requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
distro==1.9.0
idna==3.10
importlib_metadata==8.7.0
kagglehub==0.3.12
Markdown==3.8.2
numpy==2.0.2
-e git+https://github.com/octondata/od-parse.git@291daf47030dea6cde029f2e91c3483d4596e89c#egg=od_parse
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
pycparser==2.22
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
tabula-py==2.10.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
